fix(nvim): move lazygit config inside the toggleterm config to lazyload toggleterm
