--- .config/nvim/lua/lazygit.lua
@@ -1,12 +1,0 @@
-local Terminal  = require('toggleterm.terminal').Terminal
-local lazygit = Terminal:new({
-  cmd = "lazygit",
-  hidden = true,
-  direction = "float",
-})
-
-function _lazygit_toggle()
-  lazygit:toggle()
-end
-
-vim.keymap.set("n", "<leader>g", "<cmd>lua _lazygit_toggle()<CR>", {noremap = true, silent = true})
